// types.ts
export enum CustomerReturnStatus {
    Pending = 'Oczekuje na weryfikację',
    Accepted = 'Zwrot przyjęty',
}

export interface CustomerReturnItem {
    id: number;
    name: string;
    quantity: number;
    unit: string;
    reason: string;
}

export interface CustomerReturn {
    id: number;
    docNumber: string;
    returnDate: string;
    client: string;
    originalWzNumber?: string;
    items: CustomerReturnItem[];
    receivedBy: string;
    status: CustomerReturnStatus;
}